requirements.txt
bitly-api-python
pyrogram==2.0.30
requests
TgCrypto